+++ guillotina_amqp/tests/__init__.py
@@ -1,7 +1,0 @@
-import logging
-from guillotina_amqp.job import logger as logger_job
-from guillotina_amqp.worker import logger as logger_worker
-
-logging.basicConfig()
-logger_job.setLevel(10)
-logger_worker.setLevel(10)